ensure code can configure successfully
