+++ esphome/components/adc_microphone/microphone/adc_audio_microphone.h
